Moved code that exercises the Titan 3E pitch program to a test.
